Consolidation in to the AcsListener solution
