+++ AcsListener/AcsListener.sln
@@ -7,6 +7,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AcsListener", "AcsListener\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AcsListener.Tests", "AcsListener.Tests\AcsListener.Tests.csproj", "{9C6D41D5-BDC6-4E24-8381-C97C69CD71DE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RplCreator", "RplCreator\RplCreator.csproj", "{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RplCreatorTests", "RplCreatorTests\RplCreatorTests.csproj", "{B9C3249E-AAD4-448E-BC13-441A4E9E066F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedCommon", "SharedCommon\SharedCommon.csproj", "{203B029C-D970-4C88-B356-2172AA613D51}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +27,18 @@ Global
 		{9C6D41D5-BDC6-4E24-8381-C97C69CD71DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9C6D41D5-BDC6-4E24-8381-C97C69CD71DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9C6D41D5-BDC6-4E24-8381-C97C69CD71DE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B9C3249E-AAD4-448E-BC13-441A4E9E066F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B9C3249E-AAD4-448E-BC13-441A4E9E066F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B9C3249E-AAD4-448E-BC13-441A4E9E066F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B9C3249E-AAD4-448E-BC13-441A4E9E066F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{203B029C-D970-4C88-B356-2172AA613D51}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{203B029C-D970-4C88-B356-2172AA613D51}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{203B029C-D970-4C88-B356-2172AA613D51}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{203B029C-D970-4C88-B356-2172AA613D51}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
