+++ AcsListener/RplCreator/RplCreator.csproj
@@ -54,12 +54,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Program.cs" />
-    <Compile Include="ProludioCinecanvasSubtitle.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ResourcePresentationList.cs" />
-    <Compile Include="RplPlayoutId.cs" />
-    <Compile Include="RplReelDuration.cs" />
-    <Compile Include="Utf8StringWriter.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
@@ -67,6 +62,12 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="rpl_test1.xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\SharedCommon\SharedCommon.csproj">
+      <Project>{203b029c-d970-4c88-b356-2172aa613d51}</Project>
+      <Name>SharedCommon</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
