+++ AcsListener/RplCreatorTests/UnitTest1.cs
@@ -5,6 +5,8 @@ using System.Xml;
 using System.Xml.Serialization;
 using System.IO;
 
+using SharedCommon;
+
 namespace RplCreatorTests
 {
     [TestClass]
