+++ AcsListener/SharedCommon.old/Utf8StringWriter.cs
