+++ AcsListener/SharedCommon/RplReelDuration.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RplCreator
+namespace SharedCommon
 {
     public class RplReelDuration
     {
