--- RplCreator/RplCreator.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27428.2037
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RplCreator", "RplCreator\RplCreator.csproj", "{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RplCreatorTests", "RplCreatorTests\RplCreatorTests.csproj", "{B9C3249E-AAD4-448E-BC13-441A4E9E066F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0110CC8E-1E2A-466C-A1AA-0E3BAFC256DA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B9C3249E-AAD4-448E-BC13-441A4E9E066F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B9C3249E-AAD4-448E-BC13-441A4E9E066F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B9C3249E-AAD4-448E-BC13-441A4E9E066F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B9C3249E-AAD4-448E-BC13-441A4E9E066F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8C3228EE-E74E-4FB0-B5DB-9AEFBC9A2192}
-	EndGlobalSection
-EndGlobal
